more warnings

+++ libclingcon/clingcon/normalizer.h
@@ -216,7 +216,7 @@ public:
 
     std::unique_ptr< LinearPropagator > propagator_;
     bool firstRun_;
-    unsigned int varsBefore_;  /// the number of variables that we had before this
+    size_t varsBefore_;        /// the number of variables that we had before this
                                /// step
     size_t varsAfter_;         /// the highest problem specific variable in this step + 1
     size_t varsAfterFinalize_; /// the highest variable we have after this
